# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo Web para la Centralización de Documentos Electrónicos Tributarios",
        "contact": {
            "name": "Rafael Larrea",
            "email": "jrafael1108@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documento/list/recibidos/consumidor": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Obtiene una lista de documentos recibidos para el consumidor.",
                "operationId": "85b896671f2eaa51228f185255bb2983",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdUsuario": {
                                        "type": "integer"
                                    },
                                    "FechaDesde": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "FechaHasta": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/list/emitidos/empresa": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Obtiene una lista de documentos emitidos por la empresa.",
                "operationId": "72b91735621fca37efa1b77bca2d58b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdEmpresa": {
                                        "type": "integer"
                                    },
                                    "FechaDesde": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "FechaHasta": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/list/recibidos/empresa": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Obtiene una lista de documentos recibidos para la empresa.",
                "operationId": "c6003fdb9a908438cb2044c8fb5dc903",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdEmpresa": {
                                        "type": "integer"
                                    },
                                    "FechaDesde": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "FechaHasta": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/visualizar/pdf/{id}": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Visualizar PDF",
                "operationId": "10248a01eb87dce6a63b1d2f36fbca03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de transaccion, clave de acceso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la Documento"
                    },
                    "404": {
                        "description": "Documento no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/descargar/pdf/{id}": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Desacargar PDF",
                "operationId": "d87cea2ec0b282f7d596c3b5b3ca624a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de transaccion, clave de acceso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la Documento"
                    },
                    "404": {
                        "description": "Documento no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/visualizar/xml/{id}": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Obtener detalles de una Documento por ID",
                "operationId": "96d067aba523336f3b16ab4dae41fc91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de transaccion, clave de acceso ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la Documento"
                    },
                    "404": {
                        "description": "Documento no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
        "/api/documento/descargar/xml/{id}": {
            "get": {
                "tags": [
                    "Documento"
                ],
                "summary": "Desacargar xml",
                "operationId": "b9971c60dee6fe94cf0d0177d281c94f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de transaccion, clave de acceso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la Documento"
                    },
                    "404": {
                        "description": "Documento no encontrada"
                    }
                },
                "security": [
                    {
                        "Token": []
                    },
                    {
                        "Session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "description": "",
                "properties": {
                    "Success": {
                        "type": "boolean"
                    },
                    "Message": {
                        "type": "string"
                    },
                    "Data": {
                        "type": "object"
                    },
                    "Exception": {
                        "type": "object"
                    },
                    "Code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            },
            "Session": {
                "type": "apiKey",
                "description": "Enter your session id",
                "name": "SessionId",
                "in": "header"
            },
            "Apikey": {
                "type": "apiKey",
                "description": "Enter your APIKEY for company",
                "name": "APIKEY",
                "in": "header"
            }
        }
    },
...
}